+++ bricks/pretty_good_flutter_app/__brick__/android/app/src/main/kotlin/{{#pathCase}}{{org_name}}{{/pathCase}}/MainActivity.kt
@@ -1,0 +1,6 @@
+package {{#dotCase}}{{org_name}}{{/dotCase}}.{{#snakeCase}}{{project_name}}{{/snakeCase}}
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}