+++ bricks/pretty_good_flutter_app/__brick__/lib/main.dart
@@ -1,0 +1,6 @@
+import 'package:{{#snakeCase}}{{project_name}}{{/snakeCase}}/app/app.dart';
+import 'package:{{#snakeCase}}{{project_name}}{{/snakeCase}}/bootstrap.dart';
+
+void main() {
+  bootstrap(() => const App());
+}